+++ backend/canisters/user_index/api/src/queries/bot_installations.rs
@@ -2,13 +2,13 @@ use candid::CandidType;
 use oc_error_codes::OCError;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{BotInstallationLocation, BotPermissions, CanisterId, TimestampMillis};
+use types::{BotInstallationLocation, BotPermissions, TimestampMillis, UserId};
 
 #[ts_export(user_index, bot_installations)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
-    pub installed_since: Option<BotInstallationLocation>,
-    pub max_results: u16,
+    pub from: u32,
+    pub size: u16,
 }
 
 #[ts_export(user_index, bot_installations)]
@@ -21,15 +21,30 @@ pub enum Response {
 #[ts_export(user_index, bot_installations)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct SuccessResult {
-    pub installations: Vec<InstallationDetails>,
+    pub events: Vec<BotInstallationEvent>,
 }
 
 #[ts_export(user_index, bot_installations)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub struct InstallationDetails {
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub enum BotInstallationEvent {
+    Installed(BotInstalled),
+    Uninstalled(BotUninstalled),
+}
+
+#[ts_export(user_index, bot_installations)]
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct BotInstalled {
     pub location: BotInstallationLocation,
-    pub installed_at: TimestampMillis,
-    pub local_user_index: CanisterId,
     pub granted_permissions: BotPermissions,
     pub granted_autonomous_permissions: BotPermissions,
+    pub installed_by: UserId,
+    pub timestamp: TimestampMillis,
+}
+
+#[ts_export(user_index, bot_installations)]
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct BotUninstalled {
+    pub location: BotInstallationLocation,
+    pub uninstalled_by: UserId,
+    pub timestamp: TimestampMillis,
 }